Move out error view model to decouple it from signin view model.
